Restructure metadata legacy classes
